requirements.txt
certifi==2025.8.3
charset-normalizer==3.4.3
gensim==4.3.3
idna==3.10
numpy==1.26.4
pandas==2.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.5
scipy==1.13.1
six==1.17.0
smart_open==7.3.0.post1
tzdata==2025.2
urllib3==2.5.0
wrapt==1.17.3
